+++ src/app/(public)/(Home)/page.tsx
@@ -1,0 +1,9 @@
+import styles from "./page.module.css";
+
+export default function Home() {
+  return (
+    <>
+      
+    </>
+  );
+}
