+++ src/shared/styles/globals.scss
@@ -1,0 +1,11 @@
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  list-style: none;
+}
+
+.container {
+  width: 1240px;
+  margin: 0 auto;
+}
